Cleaned up some data
