+++ src/components/app.tsx
@@ -2,19 +2,16 @@ import React, { useState, useEffect } from 'react'
 import { connect, ConnectedProps } from 'react-redux'
 import { Switch, Route, Router } from 'react-router-dom'
 
-import { Token, User } from 'devu-shared-modules'
+import AuthProvider from 'components/pages/authProvider'
+import LoadingOverlay from 'components/shared/loaders/loadingOverlay'
+import HomePage from 'components/pages/homePage'
+import NotFoundPage from 'components/pages/notFoundPage'
 
 import { setUser } from 'redux/actions/user.actions'
 
-import RequestService from 'services/request.service'
 import history from 'services/history.service'
 
-import AuthProvider from 'components/pages/authProvider'
-import LoadingOverlay from 'components/shared/loaders/loadingOverlay'
-
-import HomePage from 'components/pages/homePage'
-import NotFoundPage from 'components/pages/notFoundPage'
-import { decodeAccessToken } from 'services/authentication.service'
+import fetchUser from 'utils/fetchUser.utils'
 
 const mapDispatch = { setUser }
 const connector = connect(null, mapDispatch)
@@ -26,36 +23,12 @@ const App = ({ setUser }: Props) => {
   const [error, setError] = useState(null)
   const [authenticated, setAuthenticated] = useState(false)
 
-  const getUserInfo = async () => {
-    try {
-      const loginResponse: Token = await RequestService.get('/api/login', { credentials: 'include' })
-      const decodedToken = decodeAccessToken(loginResponse.accessToken)
-
-      // Normally the RequestService will handle all of this for us
-      // But because the auth token hasn't been written to the store yet (and the RequestService attempts to get
-      // the access token from the store), we have to setup the request by hand
-      const options = {
-        headers: {
-          accept: 'application/json',
-          'content-type': 'application/json',
-          authorization: `Bearer ${loginResponse.accessToken}`,
-        },
-      }
-
-      // Telling the request service to not use it's auth (it's not there) and passing by hand via options
-      const user: User = await RequestService.get(`/api/users/${decodedToken.userId}`, options, true)
-
-      setUser({ ...loginResponse, ...user })
-      setAuthenticated(true)
-    } catch (err) {
-      setError(err)
-    } finally {
-      setLoading(false)
-    }
-  }
-
   useEffect(() => {
-    getUserInfo()
+    fetchUser()
+      .then(setUser)
+      .then(() => setAuthenticated(true))
+      .catch(setError)
+      .finally(() => setLoading(false))
   }, [])
 
   if (loading) return <LoadingOverlay delay={100} />
